Löffler published a simpler recurrence relation for the number of partitions in the Wilcox distribution: https://upload.wikimedia.org/wikipedia/commons/f/f5/LoefflerWilcoxonMannWhitneyTest.pdf

This reduces computation time and memory allocations.

## master

```julia
julia> using StatsFuns, Chairmarks

julia> @be StatsFuns.wilcoxpdf.(1, 1, 0:1)
Benchmark: 6595 samples with 169 evaluations
 min    53.006 ns (6 allocs: 272 bytes)
 median 60.160 ns (6 allocs: 272 bytes)
 mean   78.120 ns (6 allocs: 272 bytes, 0.32% gc time)
 max    6.206 μs (6 allocs: 272 bytes, 97.99% gc time)

julia> @be StatsFuns.wilcoxpdf.(10, 20, 0:200)
Benchmark: 96 samples with 1 evaluation
 min    866.500 μs (583 allocs: 1.663 MiB)
 median 931.500 μs (583 allocs: 1.663 MiB)
 mean   1.042 ms (583 allocs: 1.663 MiB, 5.01% gc time)
 max    4.602 ms (583 allocs: 1.663 MiB, 70.37% gc time)
```

## This PR

```julia
julia> using StatsFuns, Chairmarks

julia> @be StatsFuns.wilcoxpdf.(1, 1, 0:1)
Benchmark: 6210 samples with 205 evaluations
 min    48.371 ns (8 allocs: 272 bytes)
 median 51.424 ns (8 allocs: 272 bytes)
 mean   70.230 ns (8 allocs: 272 bytes, 0.19% gc time)
 max    17.490 μs (8 allocs: 272 bytes, 98.68% gc time)

julia> @be StatsFuns.wilcoxpdf.(10, 20, 0:200)
Benchmark: 451 samples with 1 evaluation
 min    178.375 μs (804 allocs: 187.562 KiB)
 median 188.667 μs (804 allocs: 187.562 KiB)
 mean   210.476 μs (804 allocs: 187.562 KiB, 0.88% gc time)
 max    4.608 ms (804 allocs: 187.562 KiB, 95.17% gc time)
```